+++ setup.cfg
@@ -33,14 +33,14 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-    check-dead-queue = monitoring.tools.check_dead_queues:cli
-    s2-gap-filler = monitoring.tools.s2_gap_filler:cli
-    s2-gap-report = monitoring.tools.s2_gap_report:cli
-    download-chirps = data.tools.chirps:cli
+    check-dead-queue = tools.monitoring.check_dead_queues:cli
+    s2-gap-filler = tools.monitoring.s2_gap_filler:cli
+    s2-gap-report = tools.monitoring.s2_gap_report:cli
+    download-chirps = tools.data.chirps:cli
 
 [options.packages.find]
 exclude =
-    tests*
+    tools.tests*
 include =
-    monitoring*
-    data*
+    tools.monitoring*
+    tools.data*
