+++ tools/tests/data/inventory/data_file.csv.gz
