+++ tools/tests/data/inventory/manifest.json
