+++ tools/tests/test_s2_gap_filler.py
@@ -1,5 +1,4 @@
 import gzip
-from pathlib import Path
 from random import randrange
 from unittest.mock import patch
 
@@ -7,16 +6,9 @@ import boto3
 from click.testing import CliRunner
 from moto import mock_s3, mock_sqs
 from odc.aws.queue import get_queue
-from urlpath import URL
 
-from monitoring.tests.conftest import (
-    FAKE_STAC_FILE,
-    REGION,
-    SQS_QUEUE_NAME,
-    TEST_BUCKET_NAME,
-    COGS_REGION,
-)
-from monitoring.tools import s2_gap_filler
+from tools.monitoring import s2_gap_filler
+from tools.tests.conftest import *
 
 
 @mock_s3
