+++ tools/tests/test_s2_gap_report.py
@@ -3,12 +3,12 @@ from unittest.mock import patch
 import boto3
 from moto import mock_s3
 
-from monitoring.tests.conftest import *
-from monitoring.tools import s2_gap_report
-from monitoring.tools.s2_gap_report import (
+from tools.monitoring import s2_gap_report
+from tools.monitoring.s2_gap_report import (
     get_and_filter_cogs_keys,
     generate_buckets_diff,
 )
+from tools.tests.conftest import *
 
 
 @mock_s3
@@ -48,7 +48,7 @@ def test_get_and_filter_cogs_keys(
     )
 
     with patch.object(
-        s2_gap_report, "SOURCE_INVENTORY_PATH", str(s3_inventory_path)
+            s2_gap_report, "SOURCE_INVENTORY_PATH", str(s3_inventory_path)
     ), patch.object(s2_gap_report, "BASE_FOLDER_NAME", str(INVENTORY_FOLDER)):
         scenes_list = get_and_filter_cogs_keys()
         assert len(scenes_list) == 6
@@ -119,7 +119,7 @@ def test_generate_buckets_diff(
     )
 
     with patch.object(
-        s2_gap_report, "SOURCE_INVENTORY_PATH", str(s3_cogs_inventory_path)
+            s2_gap_report, "SOURCE_INVENTORY_PATH", str(s3_cogs_inventory_path)
     ), patch.object(
         s2_gap_report, "SENTINEL_2_INVENTORY_PATH", str(s3_inventory_path)
     ), patch.object(
